# apps/docs/lib/engine-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "meter engine API",
    "description": "The metering engine's HTTP surface — money-truth and usage. The source of truth for engine SDK codegen.",
    "contact": {
      "name": "Josh Galbreath"
    },
    "license": {
      "name": "AGPL-3.0-only"
    },
    "version": "0.0.0"
  },
  "paths": {
...
    "/v1/accounts/{id}/events": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "`GET /v1/accounts/{id}/events`",
        "operationId": "list_for_account",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Account id (UUID)",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The account's latest (non-voided) events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Event"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/accounts/{id}/usage-by-day": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "`GET /v1/accounts/{id}/usage-by-day?start&end` — daily credit usage time series.",
        "operationId": "usage_by_day",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Account id (UUID)",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "start",
            "in": "query",
            "description": "Period start (RFC3339)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "end",
            "in": "query",
            "description": "Period end (RFC3339)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Daily credit totals over the period",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DayUsage"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/events": {
      "post": {
        "tags": [
          "events"
        ],
        "summary": "`POST /v1/events`",
        "operationId": "record",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordEventBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Event recorded (idempotent on org + key)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          }
        }
      }
    },
    "/v1/events/batch": {
      "post": {
        "tags": [
          "events"
        ],
        "summary": "`POST /v1/events/batch` — bulk ingest. Returns `202 Accepted` with the count recorded; ids are",
        "description": "content-addressed from `(org_id, idempotency_key)`, so callers can derive them without the payload.",
        "operationId": "record_batch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordBatchBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Batch accepted; returns the count recorded"
          }
        }
      }
    },
    "/v1/events/{id}": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "`GET /v1/events/{id}`",
        "operationId": "get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Event id (UUID)",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The event",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "404": {
            "description": "Unknown event"
          }
        }
      }
    },
    "/v1/events/{id}/amend": {
      "post": {
        "tags": [
          "events"
        ],
        "summary": "`POST /v1/events/{id}/amend`",
        "operationId": "amend",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Event id (UUID) to amend",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AmendBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "New event version superseding the prior",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/orgs/{id}/event-count": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "`GET /v1/orgs/{id}/event-count` — count of an organization's live (recorded) events.",
        "operationId": "event_count",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Org id (UUID)",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Count of live (recorded) events",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventCountResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/orgs/{id}/usage-by-day": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "`GET /v1/orgs/{id}/usage-by-day` — daily event + credit totals from the event store (`ClickHouse`).",
        "operationId": "org_usage_by_day",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Org id (UUID)",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Daily event + credit totals",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventDayUsage"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/orgs/{id}/usage-by-model": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "`GET /v1/orgs/{id}/usage-by-model` — usage aggregated by model, derived from the event store",
        "description": "(`ClickHouse`). Reflects amends and voids; ordered by spend, highest first.",
        "operationId": "usage_by_model",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Org id (UUID)",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usage aggregated by model, ordered by spend",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ModelUsage"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/runs/{id}/void": {
      "post": {
        "tags": [
          "events"
        ],
        "summary": "`POST /v1/runs/{id}/void` — kill a whole run. Voids the run's events (append-only: each is marked",
        "description": "voided) and reverses its ledger impact (release open holds, refund settled charges). Both halves\nare idempotent, so retrying a void is safe.",
        "operationId": "void_run",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Run id (UUID) to kill",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Counts of events voided + ledger holds/charges reversed"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AmendBody": {
        "type": "object",
        "description": "`POST /v1/events/{id}/amend`",
        "required": [
          "properties"
        ],
        "properties": {
          "properties": {
            "type": "object",
            "description": "Replacement customer JSON (the amend records a new version)."
          }
        }
      },
...
      "DayUsage": {
        "type": "object",
        "description": "Credits consumed on a single UTC day — a point in a usage time series.",
        "required": [
          "day",
          "total_credits",
          "entry_count"
        ],
        "properties": {
          "day": {
            "type": "string",
            "description": "The UTC day as `YYYY-MM-DD`."
          },
          "entry_count": {
            "type": "integer",
            "format": "int64"
          },
          "total_credits": {
            "type": "string",
            "description": "Total credits consumed that day, as an exact decimal string."
          }
        }
      },
...
      "Event": {
        "type": "object",
        "description": "One usage event. `properties` is arbitrary customer JSON; `run_id` groups the events of one agent\nrun. Amendments create a new event whose `supersedes` points back to the version it replaces.\n\n`Event` is not `Eq` because `properties` (a `serde_json::Value`) is only `PartialEq`.",
        "required": [
          "id",
          "org_id",
          "idempotency_key",
          "event_time",
          "meter",
          "account_id",
          "properties",
          "status",
          "created_at"
        ],
        "properties": {
          "account_id": {
            "type": "string",
            "format": "uuid"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "event_time": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "idempotency_key": {
            "type": "string"
          },
          "meter": {
            "type": "string"
          },
          "org_id": {
            "type": "string",
            "format": "uuid"
          },
          "properties": {
            "type": "object",
            "description": "Arbitrary customer JSON."
          },
          "run_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/EventStatus"
          },
          "supersedes": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        }
      },
      "EventCountResponse": {
        "type": "object",
        "description": "The count of an organization's live (recorded) events.",
        "required": [
          "count"
        ],
        "properties": {
          "count": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
      "EventDayUsage": {
        "type": "object",
        "description": "Daily usage totals for an organization (a time series for charts). `day` is `YYYY-MM-DD`.",
        "required": [
          "day",
          "events",
          "credits"
        ],
        "properties": {
          "credits": {
            "type": "number",
            "format": "double"
          },
          "day": {
            "type": "string"
          },
          "events": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
      "EventStatus": {
        "type": "string",
        "description": "The lifecycle status of an event version.",
        "enum": [
          "recorded",
          "amended",
          "voided"
        ]
      },
...
      "ModelUsage": {
        "type": "object",
        "description": "Usage aggregated by model for one organization.",
        "required": [
          "model",
          "events",
          "input_tokens",
          "output_tokens",
          "credits"
        ],
        "properties": {
          "credits": {
            "type": "number",
            "format": "double"
          },
          "events": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "input_tokens": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "model": {
            "type": "string"
          },
          "output_tokens": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "RecordBatchBody": {
        "type": "object",
        "description": "`POST /v1/events/batch` — the firehose ingest path: many events in one round-trip, written in a\nsingle bulk insert. Idempotent per event on `(org_id, idempotency_key)`, exactly like `record`.",
        "required": [
          "events"
        ],
        "properties": {
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecordEventBody"
            }
          }
        }
      },
      "RecordEventBody": {
        "type": "object",
        "description": "`POST /v1/events`",
        "required": [
          "org_id",
          "idempotency_key",
          "meter",
          "account"
        ],
        "properties": {
          "account": {
            "type": "string",
            "format": "uuid"
          },
          "event_time": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "idempotency_key": {
            "type": "string"
          },
          "meter": {
            "type": "string"
          },
          "org_id": {
            "type": "string",
            "format": "uuid"
          },
          "properties": {
            "type": "object",
            "description": "Arbitrary customer JSON describing the usage."
          },
          "run_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        }
      },
...
    }
  },
...
}